+++ lib/is-primitive.js
@@ -1,0 +1,3 @@
+module.exports = function isPrimitive (value) {
+  return value === null || typeof value !== 'object'
+}
