+++ lib/same-regexp.js
@@ -1,0 +1,3 @@
+module.exports = function sameRegExp (a, b) {
+  return a.source === b.source && a.flags === b.flags
+}
